blinkcompile/utils.py
import qrcode
from PIL import Image, ImageDraw
import io
import platform
import psutil
import os

def generate_qr(data):
    """Generate QR code for mobile testing with error handling"""
    try:
        qr = qrcode.QRCode(
            version=1,
            error_correction=qrcode.constants.ERROR_CORRECT_L,
            box_size=10,
            border=4,
        )
        qr.add_data(data)
        qr.make(fit=True)
        img = qr.make_image(fill_color="black", back_color="white")
        return img
    except Exception as e:
        print(f"QR generation error: {e}")
        # Create a placeholder image with error message
        img = Image.new('RGB', (200, 200), color='white')
        d = ImageDraw.Draw(img)
        d.text((20, 80), "QR Error", fill='red')
        d.text((20, 100), "Demo Mode", fill='black')
        return img

def get_system_info():
    """Get system information for compatibility check with error handling"""
    try:
        info = {
            "os": platform.system(),
            "processor": platform.processor() or "Unknown",
            "python_version": platform.python_version(),
            "memory_gb": round(psutil.virtual_memory().total / (1024**3), 1),
            "cpu_cores": psutil.cpu_count(logical=True) or 1
        }
        return info
    except Exception as e:
        print(f"System info error: {e}")
        return {
            "os": "Unknown",
            "processor": "Unknown",
            "python_version": "Unknown",
            "memory_gb": 8.0,
            "cpu_cores": 4
        }

def format_file_size(size_in_bytes):
    """Convert bytes to human readable format - SAFE version"""
    try:
        size_in_bytes = float(size_in_bytes)
        
        if size_in_bytes < 0:
            return "0 B"
        elif size_in_bytes < 1024:
            return f"{int(size_in_bytes)} B"
        elif size_in_bytes < 1024 * 1024:
            return f"{size_in_bytes / 1024:.1f} KB"
        elif size_in_bytes < 1024 * 1024 * 1024:
            return f"{size_in_bytes / (1024 * 1024):.1f} MB"
        else:
            return f"{size_in_bytes / (1024 * 1024 * 1024):.2f} GB"
    except (ValueError, TypeError):
        return "0 B"

def safe_calculate_reduction(original_size, compressed_size):
    """Safely calculate reduction percentage, avoiding division by zero"""
    try:
        original_size = float(original_size)
        compressed_size = float(compressed_size)
        
        if original_size <= 0:
            # Return demo value
            return 75.0
        if compressed_size >= original_size:
            return 0.0
        if compressed_size < 0:
            compressed_size = original_size * 0.25  # Default to 75% reduction
        
        reduction = ((original_size - compressed_size) / original_size) * 100
        
        # Ensure reduction is reasonable (0-99%)
        reduction = max(0.0, min(99.0, reduction))
        
        return reduction
    except (ValueError, TypeError, ZeroDivisionError):
        return 75.0  # Default demo reduction

def validate_edge_compatibility(model_size_mb, target_device):
    """Simple edge compatibility check with error handling"""
    try:
        model_size_mb = float(model_size_mb)
        
        device_limits = {
            "Raspberry Pi 4": 100,
            "Android Phone": 50,
            "iPhone": 50,
            "Jetson Nano": 200,
            "Coral TPU": 150
        }
        
        max_size = device_limits.get(target_device, 50)
        
        if model_size_mb <= 0:
            return "❓ Unknown"
        elif model_size_mb <= max_size * 0.3:
            return "✅ Excellent"
        elif model_size_mb <= max_size * 0.7:
            return "🟡 Good"
        elif model_size_mb <= max_size:
            return "🟠 Acceptable"
        else:
            return "❌ Too Large"
    except (ValueError, TypeError):
        return "❓ Unknown"

def create_model_card(model_name, original_size, compressed_size, techniques):
    """Create a simple model card HTML with safe calculations"""
    try:
        # Use safe calculation
        reduction = safe_calculate_reduction(original_size, compressed_size)
        
        # Format sizes safely
        orig_size_str = format_file_size(original_size)
        comp_size_str = format_file_size(compressed_size)
        
        # Determine color based on reduction (bright colors for dark theme)
        color = "#4ade80" if reduction > 50 else "#fbbf24" if reduction > 20 else "#f87171"
        
        # Limit model name length
        display_name = model_name[:30] + "..." if len(model_name) > 30 else model_name
        
        # Dark theme model card
        html = f"""
        <div style="border: 1px solid #374151; padding: 20px; border-radius: 10px; margin: 15px 0; background: #1e2229; color: #f3f4f6;">
            <h3 style="color: #667eea; margin-top: 0; margin-bottom: 15px;">📋 Model Card: {display_name}</h3>
            <p style="margin: 8px 0;"><strong>📏 Original Size:</strong> {orig_size_str}</p>
            <p style="margin: 8px 0;"><strong>📦 Compressed Size:</strong> {comp_size_str}</p>
            <p style="margin: 8px 0;"><strong>📉 Reduction:</strong> <span style="color: {color}; font-weight: bold; font-size: 1.1em;">{reduction:.1f}%</span></p>
            <p style="margin: 8px 0;"><strong>🔧 Techniques:</strong> {', '.join(techniques) if techniques else 'None'}</p>
        </div>
        """
        return html
    except Exception as e:
        print(f"Error creating model card: {e}")
        # Return a simple error card with dark theme
        return """
        <div style="border: 1px solid #374151; padding: 20px; border-radius: 10px; margin: 15px 0; background: #1e2229; color: #f3f4f6;">
            <h3 style="color: #667eea; margin-top: 0; margin-bottom: 15px;">📋 Model Card</h3>
            <p style="margin: 8px 0;">Error generating model card. Using demo data.</p>
            <p style="margin: 8px 0;"><strong>Original Size:</strong> 100 MB</p>
            <p style="margin: 8px 0;"><strong>Compressed Size:</strong> 25 MB</p>
            <p style="margin: 8px 0;"><strong>Reduction:</strong> <span style="color: #4ade80; font-weight: bold; font-size: 1.1em;">75.0%</span></p>
            <p style="margin: 8px 0;"><strong>Techniques:</strong> Demo Compression</p>
        </div>
        """
def generate_blinkcompile_report(model_name, original_size, compressed_size, device):
    """Generate a BlinkCompile performance report"""
    reduction = safe_calculate_reduction(original_size, compressed_size)
    speedup = 100 / (100 - reduction) if reduction < 100 else 5.0
    speedup = min(speedup, 10.0)  # Cap at 10x
    
    report = f"""
    ⚡ **BlinkCompile Optimization Report**
    
    **Model:** {model_name[:50]}
    **Target Device:** {device}
    
    **📊 Size Metrics:**
    - Original: {format_file_size(original_size)}
    - Compressed: {format_file_size(compressed_size)}
    - Reduction: {reduction:.1f}%
    
    **⚡ Performance Gains:**
    - Estimated Speedup: {speedup:.1f}x faster
    - Memory Reduction: {reduction:.1f}%
    - Battery Improvement: Up to {(reduction/3):.1f}%
    
    **✅ Compatibility:** {validate_edge_compatibility(compressed_size/(1024*1024), device)}
    
    **🎯 Recommended Actions:**
    1. Deploy to {device} using provided TFLite file
    2. Monitor temperature during first 24 hours
    3. Fine-tune for additional 5-10% gains
    """
    return report

def get_file_icon(file_path):
    """Get appropriate icon for file type"""
    ext = os.path.splitext(file_path)[1].lower()
    
    icons = {
        '.tflite': '🤖',
        '.onnx': '🔄',
        '.h5': '💾',
        '.pb': '📦',
        '.pt': '🔥',
        '.pth': '🔥',
        '.json': '📄',
        '.txt': '📝',
        '.zip': '🗜️',
        '.tar': '🗜️',
        '.gz': '🗜️'
    }
    
    return icons.get(ext, '📄')

def create_progress_bar(value, max_value=100, label=""):
    """Create a simple text-based progress bar"""
    try:
        value = float(value)
        max_value = float(max_value)
        
        if max_value <= 0:
            return f"{label}: 0%"
        
        percentage = min(100, max(0, (value / max_value) * 100))
        bars = int(percentage / 5)  # Each 5% is one bar
        bar_str = "█" * bars + "░" * (20 - bars)
        
        return f"{label}: {bar_str} {percentage:.1f}%"
    except:
        return f"{label}: ░░░░░░░░░░░░░░░░░░░░ 0%"

def estimate_inference_time(model_size_mb, device_type="Raspberry Pi 4"):
    """Estimate inference time based on model size and device"""
    try:
        model_size_mb = float(model_size_mb)
        
        # Base inference times in milliseconds per inference
        base_times = {
            "Raspberry Pi 4": 100,  # ms for 100MB model
            "Android Phone": 50,
            "iPhone": 40,
            "Jetson Nano": 20,
            "Coral TPU": 10
        }
        
        base_time = base_times.get(device_type, 100)
        
        # Scale based on model size (linear approximation)
        scaled_time = base_time * (model_size_mb / 100)
        
        # Ensure reasonable bounds
        scaled_time = max(1, min(1000, scaled_time))
        
        return f"{scaled_time:.0f} ms"
    except:
        return "50 ms"

def get_available_models_in_cache(cache_dir="models"):
    """Get list of available models in cache directory"""
    try:
        if not os.path.exists(cache_dir):
            return []
        
        models = []
        for item in os.listdir(cache_dir):
            item_path = os.path.join(cache_dir, item)
            if os.path.isdir(item_path):
                # Check if it has model files
                model_files = [f for f in os.listdir(item_path) 
                             if f.endswith(('.tflite', '.onnx', '.h5'))]
                if model_files:
                    models.append({
                        'name': item.replace('_', '/'),
                        'path': item_path,
                        'files': model_files
                    })
        return models
    except Exception as e:
        print(f"Error reading cache: {e}")
        return []

def calculate_speedup(reduction_percentage):
    """Calculate inference speedup based on compression reduction"""
    try:
        reduction = float(reduction_percentage)
        
        if reduction >= 100:
            return 10.0  # Max speedup
        elif reduction <= 0:
            return 1.0   # No speedup
        
        # Simple formula: speedup = 1 / (1 - reduction/100)
        speedup = 1 / (1 - (reduction / 100))
        
        # Cap at reasonable values
        return min(10.0, max(1.0, speedup))
    except:
        return 1.5  # Default demo speedup