blinkcompile/requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
transformers==4.35.0
tf2onnx==1.15.0
onnx==1.14.0
onnxruntime==1.16.0
qrcode[pil]==7.4.0
Pillow==10.0.0
numpy==1.24.0
protobuf==3.20.0
psutil==5.9.6